Moved code from compute/exec to compute to reduce entanglement with acero
