+++ cpp/src/arrow/compute/kernels/hash_aggregate.cc
@@ -30,7 +30,7 @@
 #include "arrow/compute/api_aggregate.h"
 #include "arrow/compute/api_vector.h"
 #include "arrow/compute/key_hash.h"
-#include "arrow/compute/exec/key_map.h"
+#include "arrow/compute/key_map.h"
 #include "arrow/compute/exec/util.h"
 #include "arrow/compute/exec_internal.h"
 #include "arrow/compute/kernel.h"
