+++ cpp/src/arrow/compute/row/compare_internal.cc
@@ -22,7 +22,9 @@
 #include <algorithm>
 #include <cstdint>
 
-#include "arrow/compute/exec/util.h"
+// DIPO
+#include "arrow/compute/util.h"
+#include "arrow/compute/util_internal.h"
 #include "arrow/util/bit_util.h"
 #include "arrow/util/ubsan.h"
