+++ cpp/src/arrow/compute/row/row_internal.cc
@@ -17,7 +17,8 @@
 
 #include "arrow/compute/row/row_internal.h"
 
-#include "arrow/compute/exec/util.h"
+// DIPO
+#include "arrow/compute/util.h"
 
 namespace arrow {
 namespace compute {
